projects/arbitraje/pyproject.toml
[tool.poetry]
name = "arbitraje"
version = "1.4.1"
description = "Crypto price arbitrage scanner using ccxt (BF/tri modes)"
authors = ["Your Name <you@example.com>"]
license = "MIT"

packages = [
  { include = "arbitraje", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
requests = "^2.32.0"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"
ccxt = "^4.4.23"
binance-sdk-spot = "^3.0.0"
PyYAML = "^6.0.2"
ujson = "^5.10.0"
numpy = "^2.3.4"
numba = "^0.62.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
ruff = "^0.12.2"

[tool.ruff]
line-length = 120
# Exclude generated or vendor SDKs and large test data directories from ruff in this subproject
exclude = [
  "sdk/**",
  "sdk/okx/**",
  "artifacts/**",
  "tests/**/data/**",
  "**/*.ipynb",
  "vendor/**",
  "third_party/**",
]

[tool.ruff.lint]
# moved select/ignore settings into the lint subsection to avoid deprecation warnings
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.poetry.scripts]
arbitraje-ccxt = "arbitraje.arbitrage_report_ccxt:main"
arbitraje-tri = "arbitraje.tri_bot:main"
swapper = "arbitraje.swapper:_main_cli"
swaper = "arbitraje.swapper:_main_cli"  # backward-compat command name

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.dataqbs]
# Semantic engine label for releases (human-friendly)
engine = "3.2"
